+++ shell/src/main/java/de/ollie/carp/bm/shell/command/BattleMapTokenCommands.java
@@ -26,8 +26,8 @@ public class BattleMapTokenCommands {
 			value = "tokenIdOrName"
 		) String tokenIdOrName,
 		@ShellOption(help = "Id or name of the battle map.", value = "battleMapIdOrName") String battleMapIdOrName,
-		@ShellOption(help = "X coordinate where the token is to place.", value = "x") int x,
-		@ShellOption(help = "Y coordinate where the token is to place.", value = "y") int y
+		@ShellOption(help = "X coordinate where the token is to place.", value = "x") double x,
+		@ShellOption(help = "Y coordinate where the token is to place.", value = "y") double y
 	) {
 		try {
 			return client.setTokenToBattleMapOfSpielrunde(
